ct: Split ct/level_zero/reader
Extract top level L0 reader into its own module.
Currently, the 'reader' module contains both pipeline component that
fetches data for the read path and the actual L0 read path (log reader).
Thie reader is used from the kafka/data layer through the ct/frontend
module. This commit fixes this by extracting top level L0 reader into
the fe_reader module.
