+++ src/v/cloud_topics/level_zero/frontend_reader/reader.h
