+++ src/v/cloud_topics/level_zero/stm/BUILD
@@ -13,6 +13,7 @@ redpanda_cc_library(
     visibility = [
         "//src/v/cloud_topics/batch_cache:__pkg__",
         "//src/v/cloud_topics/frontend:__pkg__",
+        "//src/v/cloud_topics/level_zero/frontend_reader:__pkg__",
         "//src/v/cloud_topics/level_zero/reader:__pkg__",
         "//src/v/cloud_topics/level_zero/reader/tests:__pkg__",
     ],
